Implement error counting
Special care is taken here to prevent our error counter from
decrementing if a pipeline that we once observed to fail is retried
and then succeeds.

A decremented prometheus counter is interpreted as a reset by the
prometheus server, and that's not want we want.  Observed errors are
errors, even after a pipeline is retried and fixed.